core/python/requirements.txt
pandas==2.2.3
numpy==2.1.2
psycopg2-binary>=2.9.10
sqlalchemy==2.0.36
scikit-learn==1.5.2
